Working Android and iOS sign in and sign up
